+++ docs/CHANGELOG.md
@@ -18,5 +18,6 @@
 + Move all OSGI dependencies to h2gis-functions-osgi and to postgis-jts-osgi
 + Move classes used for datasource creation (`H2GISOsgiDBFactory` and `DataSourceFactoryImpl`) to `*-osgi` package and keep test purpose class (`HGISSimpleDBFactory`) in `h2gis-functions`
 + Fix `ST_OrderingEquals` and `ST_AsBinary` parameters to handle the case of a null geometry
++ Fix ST_Accum and ST_Union to filter empty geometry
 + Fix `FindGeometryMetadata` to handle the case of case-sensitive names
 + Make the method `PostGISDBFactory.createDataSource(Properties)` static